+++ adevspoon-domain/src/main/kotlin/com/adevspoon/domain/techQuestion/dto/request/GetQuestionAnswerList.kt
@@ -6,6 +6,6 @@ data class GetQuestionAnswerList(
     val memberId: Long,
     val questionId: Long?,
     val sort: QuestionAnswerListSortType,
-    val offset: Int = 0,
+    val offset: Long = 0,
     val limit: Int = 10,
 )
